+++ crates/core/executor/src/events/syscall.rs
@@ -10,9 +10,9 @@ use crate::syscalls::SyscallCode;
 #[repr(C)]
 pub struct SyscallEvent {
     /// The program counter.
-    pub pc_rel: u64,
+    pub pc_rel: u32,
     /// The next program counter.
-    pub next_pc_rel: u64,
+    pub next_pc_rel: u32,
     /// The shard number.
     pub shard: u32,
     /// The clock cycle.
